+++ src/server.c
@@ -6660,7 +6660,6 @@ struct serverTest {
     int failed;
 } serverTests[] = {
     {"quicklist", quicklistTest},
-    {"zipmap", zipmapTest},
     {"dict", dictTest},
     {"listpack", listpackTest},
 };
