+++ src/zipmap.c
@@ -432,101 +432,3 @@ int zipmapValidateIntegrity(unsigned char *zm, size_t size, int deep) {
     return 1;
 #undef OUT_OF_RANGE
 }
-
-#ifdef SERVER_TEST
-static void zipmapRepr(unsigned char *p) {
-    unsigned int l;
-
-    printf("{status %u}", *p++);
-    while (1) {
-        if (p[0] == ZIPMAP_END) {
-            printf("{end}");
-            break;
-        } else {
-            unsigned char e;
-
-            l = zipmapDecodeLength(p);
-            printf("{key %u}", l);
-            p += zipmapEncodeLength(NULL, l);
-            if (l != 0 && fwrite(p, l, 1, stdout) == 0) perror("fwrite");
-            p += l;
-
-            l = zipmapDecodeLength(p);
-            printf("{value %u}", l);
-            p += zipmapEncodeLength(NULL, l);
-            e = *p++;
-            if (l != 0 && fwrite(p, l, 1, stdout) == 0) perror("fwrite");
-            p += l + e;
-            if (e) {
-                printf("[");
-                while (e--) printf(".");
-                printf("]");
-            }
-        }
-    }
-    printf("\n");
-}
-
-#define UNUSED(x) (void)(x)
-int zipmapTest(int argc, char *argv[], int flags) {
-    unsigned char *zm;
-
-    UNUSED(argc);
-    UNUSED(argv);
-    UNUSED(flags);
-
-    zm = zipmapNew();
-
-    zm = zipmapSet(zm, (unsigned char *)"name", 4, (unsigned char *)"foo", 3, NULL);
-    zm = zipmapSet(zm, (unsigned char *)"surname", 7, (unsigned char *)"foo", 3, NULL);
-    zm = zipmapSet(zm, (unsigned char *)"age", 3, (unsigned char *)"foo", 3, NULL);
-    zipmapRepr(zm);
-
-    zm = zipmapSet(zm, (unsigned char *)"hello", 5, (unsigned char *)"world!", 6, NULL);
-    zm = zipmapSet(zm, (unsigned char *)"foo", 3, (unsigned char *)"bar", 3, NULL);
-    zm = zipmapSet(zm, (unsigned char *)"foo", 3, (unsigned char *)"!", 1, NULL);
-    zipmapRepr(zm);
-    zm = zipmapSet(zm, (unsigned char *)"foo", 3, (unsigned char *)"12345", 5, NULL);
-    zipmapRepr(zm);
-    zm = zipmapSet(zm, (unsigned char *)"new", 3, (unsigned char *)"xx", 2, NULL);
-    zm = zipmapSet(zm, (unsigned char *)"noval", 5, (unsigned char *)"", 0, NULL);
-    zipmapRepr(zm);
-    zm = zipmapDel(zm, (unsigned char *)"new", 3, NULL);
-    zipmapRepr(zm);
-
-    printf("\nLook up large key:\n");
-    {
-        unsigned char buf[512];
-        unsigned char *value;
-        unsigned int vlen, i;
-        for (i = 0; i < 512; i++) buf[i] = 'a';
-
-        zm = zipmapSet(zm, buf, 512, (unsigned char *)"long", 4, NULL);
-        if (zipmapGet(zm, buf, 512, &value, &vlen)) {
-            printf("  <long key> is associated to the %d bytes value: %.*s\n", vlen, vlen, value);
-        }
-    }
-
-    printf("\nPerform a direct lookup:\n");
-    {
-        unsigned char *value;
-        unsigned int vlen;
-
-        if (zipmapGet(zm, (unsigned char *)"foo", 3, &value, &vlen)) {
-            printf("  foo is associated to the %d bytes value: %.*s\n", vlen, vlen, value);
-        }
-    }
-    printf("\nIterate through elements:\n");
-    {
-        unsigned char *i = zipmapRewind(zm);
-        unsigned char *key, *value;
-        unsigned int klen, vlen;
-
-        while ((i = zipmapNext(i, &key, &klen, &value, &vlen)) != NULL) {
-            printf("  %d:%.*s => %d:%.*s\n", klen, klen, key, vlen, vlen, value);
-        }
-    }
-    zfree(zm);
-    return 0;
-}
-#endif
