+++ src/zipmap.h
@@ -46,11 +46,6 @@ int zipmapGet(unsigned char *zm, unsigned char *key, unsigned int klen, unsigned
 int zipmapExists(unsigned char *zm, unsigned char *key, unsigned int klen);
 unsigned int zipmapLen(unsigned char *zm);
 size_t zipmapBlobLen(unsigned char *zm);
-void zipmapRepr(unsigned char *p);
 int zipmapValidateIntegrity(unsigned char *zm, size_t size, int deep);
-
-#ifdef SERVER_TEST
-int zipmapTest(int argc, char *argv[], int flags);
-#endif
 
 #endif
